+++ integration-tests/emulated/tests/assets/asset-hub-kusama/src/tests/teleport.rs
@@ -86,7 +86,6 @@ fn penpal_to_ah_foreign_assets_receiver_assertions(t: ParaToSystemParaTest) {
 	);
 	let (expected_foreign_asset_id, expected_foreign_asset_amount) =
 		non_fee_asset(&t.args.assets, t.args.fee_asset_item as usize).unwrap();
-	let expected_foreign_asset_id_v3: v3::Location = expected_foreign_asset_id.try_into().unwrap();
 
 	AssetHubKusama::assert_xcmp_queue_success(None);
 	assert_expected_events!(
@@ -103,7 +102,7 @@ fn penpal_to_ah_foreign_assets_receiver_assertions(t: ParaToSystemParaTest) {
 				who: *who == t.receiver.account_id,
 			},
 			RuntimeEvent::ForeignAssets(pallet_assets::Event::Issued { asset_id, owner, amount }) => {
-				asset_id: *asset_id == expected_foreign_asset_id_v3,
+				asset_id: *asset_id == expected_foreign_asset_id,
 				owner: *owner == t.receiver.account_id,
 				amount: *amount == expected_foreign_asset_amount,
 			},
@@ -117,7 +116,6 @@ fn ah_to_penpal_foreign_assets_sender_assertions(t: SystemParaToParaTest) {
 	AssetHubKusama::assert_xcm_pallet_attempted_complete(None);
 	let (expected_foreign_asset_id, expected_foreign_asset_amount) =
 		non_fee_asset(&t.args.assets, t.args.fee_asset_item as usize).unwrap();
-	let expected_foreign_asset_id_v3: v3::Location = expected_foreign_asset_id.try_into().unwrap();
 	assert_expected_events!(
 		AssetHubKusama,
 		vec![
@@ -133,7 +131,7 @@ fn ah_to_penpal_foreign_assets_sender_assertions(t: SystemParaToParaTest) {
 			},
 			// foreign asset is burned locally as part of teleportation
 			RuntimeEvent::ForeignAssets(pallet_assets::Event::Burned { asset_id, owner, balance }) => {
-				asset_id: *asset_id == expected_foreign_asset_id_v3,
+				asset_id: *asset_id == expected_foreign_asset_id,
 				owner: *owner == t.sender.account_id,
 				balance: *balance == expected_foreign_asset_amount,
 			},
@@ -377,7 +375,7 @@ pub fn do_bidirectional_teleport_foreign_assets_between_para_and_asset_hub_using
 	let ah_receiver_assets_before = AssetHubKusama::execute_with(|| {
 		type Assets = <AssetHubKusama as AssetHubKusamaPallet>::ForeignAssets;
 		<Assets as Inspect<_>>::balance(
-			foreign_asset_at_asset_hub_kusama.clone().try_into().unwrap(),
+			foreign_asset_at_asset_hub_kusama.clone(),
 			&AssetHubKusamaReceiver::get(),
 		)
 	});
@@ -404,7 +402,7 @@ pub fn do_bidirectional_teleport_foreign_assets_between_para_and_asset_hub_using
 	let ah_receiver_assets_after = AssetHubKusama::execute_with(|| {
 		type Assets = <AssetHubKusama as AssetHubKusamaPallet>::ForeignAssets;
 		<Assets as Inspect<_>>::balance(
-			foreign_asset_at_asset_hub_kusama.clone().try_into().unwrap(),
+			foreign_asset_at_asset_hub_kusama.clone(),
 			&AssetHubKusamaReceiver::get(),
 		)
 	});
@@ -432,7 +430,7 @@ pub fn do_bidirectional_teleport_foreign_assets_between_para_and_asset_hub_using
 		type ForeignAssets = <AssetHubKusama as AssetHubKusamaPallet>::ForeignAssets;
 		assert_ok!(ForeignAssets::transfer(
 			<AssetHubKusama as Chain>::RuntimeOrigin::signed(AssetHubKusamaReceiver::get()),
-			foreign_asset_at_asset_hub_kusama.clone().try_into().unwrap(),
+			foreign_asset_at_asset_hub_kusama.clone(),
 			AssetHubKusamaSender::get().into(),
 			asset_amount_to_send,
 		));
@@ -478,7 +476,7 @@ pub fn do_bidirectional_teleport_foreign_assets_between_para_and_asset_hub_using
 	let ah_sender_assets_before = AssetHubKusama::execute_with(|| {
 		type ForeignAssets = <AssetHubKusama as AssetHubKusamaPallet>::ForeignAssets;
 		<ForeignAssets as Inspect<_>>::balance(
-			foreign_asset_at_asset_hub_kusama.clone().try_into().unwrap(),
+			foreign_asset_at_asset_hub_kusama.clone(),
 			&AssetHubKusamaSender::get(),
 		)
 	});
@@ -504,7 +502,7 @@ pub fn do_bidirectional_teleport_foreign_assets_between_para_and_asset_hub_using
 	let ah_sender_assets_after = AssetHubKusama::execute_with(|| {
 		type ForeignAssets = <AssetHubKusama as AssetHubKusamaPallet>::ForeignAssets;
 		<ForeignAssets as Inspect<_>>::balance(
-			foreign_asset_at_asset_hub_kusama.try_into().unwrap(),
+			foreign_asset_at_asset_hub_kusama,
 			&AssetHubKusamaSender::get(),
 		)
 	});